refactor: restructure package structure
- Move TransactionCalculator to calculator/
- Move ui/transaction/ to ui/feature/transaction/
