+++ src/main/java/edu/ntnu/idatt2003/group14/ui/app/AppNavigatorImpl.java
@@ -12,8 +12,8 @@ import edu.ntnu.idatt2003.group14.ui.features.menu.options.OptionsView;
 import edu.ntnu.idatt2003.group14.ui.features.portfolio.GameLayout;
 import edu.ntnu.idatt2003.group14.ui.features.portfolio.PortfolioController;
 import edu.ntnu.idatt2003.group14.ui.features.portfolio.PortfolioView;
-import edu.ntnu.idatt2003.group14.ui.transaction.TransactionArchiveController;
-import edu.ntnu.idatt2003.group14.ui.transaction.TransactionArchiveView;
+import edu.ntnu.idatt2003.group14.ui.features.transaction.TransactionArchiveController;
+import edu.ntnu.idatt2003.group14.ui.features.transaction.TransactionArchiveView;
 import java.util.Objects;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
